scripts/requirements.txt
# Dependências Python para o Evolution Manager
python-dotenv>=1.0.0
